+++ src/main/java/chess/piece/PiecesGenrator.java
@@ -1,0 +1,7 @@
+package chess.piece;
+
+import java.util.List;
+
+public interface PiecesGenrator {
+    List<Piece> generate();
+}
